requirements.txt
flask>=3.0
python-dotenv>=1.0
openai>=1.60
anthropic>=0.40
google-generativeai>=0.8
requests>=2.31
